requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu118
torch==2.0.1
--extra-index-url https://download.pytorch.org/whl/cu118
torchvision==0.15.2
opencv-python
tqdm
timm
prettytable
scipy
scikit-image
kornia
numpy<2.0.0
Pillow
